12/7/2018@0530
I'm trying to figure out why the timer interrupt is taking so long, and why it is making the robot stop. Also, I can't get the OC2 interrupt to work properly. Every time I enable it, it breaks the code. I am clearing the flag bit every time the interrupt is activated. (Remember to change the timer back to The values that are currently commented out)